services/market/pyproject.toml
[project]
name = "market"
version = "0.1.0"
description = "Market data service for the Quantisti options trading platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "psycopg2-binary>=2.9.9",
    "scipy>=1.11.4",
    "numpy>=1.26.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
